pyproject.toml
[project]
name = "fastapi-iris-predictor"
version = "0.1.0"
description = "Simple API built with FastAPI for real-time predictions on the iris flower dataset using a Logistic Regression model."
authors = [
    {name = "Luís Fernando Torres",email = "luuisotorres@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uvicorn (>=0.34.2,<0.35.0)"
]

[tool.poetry]
packages = [
    {include = "fastapi_iris_predictor", from = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
